+++ src/handleTask.js
@@ -26,14 +26,4 @@ function Task(title, description, dueDate, priority) {
   this.priority = priority
 }
 
-// let task1 = new Task("Test Task1", "Description about Test", "Today","p1");
-// let task2 = new Task("Test Task2", "Description about Test", "Today","p2");
-
-// let newhandle = new handleTask();
-
-// newhandle.addTask(task1);
-// newhandle.addTask(task2);
-
-// console.log(taskStore);
-
 export default Task;
