+++ src/project.js
@@ -1,8 +1,6 @@
-
 function Project(name) {
   this.name = name
   this.todos = [];
-
 }
 
 export default Project;
